src/api/auth.py

import jwt
from datetime import datetime, timedelta

class AuthManager:
    def __init__(self):
        self.secret_key = "your-secret-key"
        self.algorithm = "HS256"

    def generate_token(self, user_id):
        # Implementation details here
        pass

    def validate_token(self, token):
        # Implementation details here
        pass

    def authenticate(self, request):
        # Implementation details here
        pass
