requirements.txt
Flask==2.1.0
Flask-RESTful==0.3.9
SQLAlchemy==1.4.31
requests==2.27.1
PyJWT==2.3.0
PyYAML==6.0
schedule==1.1.0
nltk==3.6.7
pytest==7.1.1